removing all dependencies for comparison tasks - pearson correlations are called for any instance of image save, and all instances of pearson calculations check for transform existance, and compete to see who can generate it first! A check query must be made to see if the Comparison exists in the database, otherwise there is an integrity error.

+++ neurovault/apps/statmaps/views.py
@@ -22,7 +22,6 @@ from rest_framework.renderers import JSONRenderer
 from voxel_query_functions import *
 from pybraincompare.compare import scatterplot, search
 from pybraincompare.mr.datasets import get_mni_atlas
-from neurovault.apps.statmaps.tasks import run_transformation_tasks
 from glob import glob
 import pandas
 import zipfile
@@ -418,10 +417,6 @@ def upload_folder(request, collection_cid):
                                 niftiFiles.extend(split_afni4D_to_3D(nii))
                             else:
                                 niftiFiles.append((fname,nii_path))
-
-                # We will keep track of new image objects to submit celery tasks after 
-                # to generate image transformations BEFORE pearson comparisons
-                pks_for_celery = []
                 
                 for label,fpath in niftiFiles:
                     # Read nifti file information
@@ -477,11 +472,6 @@ def upload_folder(request, collection_cid):
 
                     new_image.file = f
                     new_image.save()
-                    pks_for_celery.append(new_image.pk)
-
-                # Generate transformations and pearson calculations
-                if len(pks_for_celery) > 0 and collection.private == False:
-                    run_transformation_tasks.apply_async([pks_for_celery])
 
             except:
                 raise
@@ -507,10 +497,12 @@ def delete_image(request, pk):
     if not owner_or_contrib(request,image.collection):
         return HttpResponseForbidden()
     # Delete transformation and thumbnail
-    if os.path.exists(image.thumbnail):
-      os.remove(image.thumbnail)
-    if os.path.exists(image.transform):
-      os.remove(image.transform)
+    if image.thumbnail is not None:
+        if os.path.exists(image.thumbnail):
+            os.remove(image.thumbnail)
+    if image.transform is not None:
+        if os.path.exists(image.transform):
+            os.remove(image.transform)
     image.delete()
     return redirect('collection_details', cid=cid)
